requirements.txt
asgiref==3.8.1
Django==5.1.4
django-rest-framework==0.1.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
django-unfold==0.43.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
tzdata==2024.2
